Vocab-Builder/requirements.txt
gradio>=4.0.0
groq
python-dotenv